separate the move model
